codebase tidyup, moved helper files to services, db helper to db/repositories and removed extra routes folder
