+++ app/src/main/kotlin/diettracker/Routing.kt
@@ -1,17 +1,37 @@
 package diettracker
 
+import diettracker.db.repositories.getClientsForProfessional
+import diettracker.db.repositories.getUserIdByEmail
+import diettracker.db.repositories.getUserRoles
 import diettracker.db.tables.Clients
 import diettracker.db.tables.Users
-import diettracker.routes.quizRoutes
 import diettracker.routing.configureClientDashRoute
 import diettracker.routing.configureClientProfessionalRoutes
 import diettracker.routing.configureFoodRoutes
 import diettracker.routing.configureMessageRoutes
 import diettracker.routing.configureRecipeRoutes
+import diettracker.routing.configureStatic
 import diettracker.routing.foodDiaryRoutes
 import diettracker.routing.hasRole
 import diettracker.routing.professionalProfileRoutes
 import diettracker.routing.profileRoutes
+import diettracker.routing.quizRoutes
+import diettracker.services.ClientDietTrend
+import diettracker.services.UserSession
+import diettracker.services.buildNavbarContext
+import diettracker.services.dashboardPage
+import diettracker.services.handleViewClientDetails
+import diettracker.services.loginPage
+import diettracker.services.loginProfessional
+import diettracker.services.loginUser
+import diettracker.services.logout
+import diettracker.services.profLoginPage
+import diettracker.services.profQuizPage
+import diettracker.services.profSignUpPage
+import diettracker.services.signUpPage
+import diettracker.services.signUpProfessional
+import diettracker.services.signUpUser
+import diettracker.services.submitProfQuiz
 import io.ktor.server.application.Application
 import io.ktor.server.auth.authenticate
 import io.ktor.server.pebble.PebbleContent
