+++ app/src/main/kotlin/diettracker/routing/FoodRoutes.kt
@@ -1,17 +1,17 @@
 package diettracker.routing
 
-import diettracker.CaloriesSession
-import diettracker.UserSession
-import diettracker.addSavedMealToLog
-import diettracker.foodLogCustom
-import diettracker.foodLogRecipe
-import diettracker.foodLogReset
-import diettracker.getSavedMeals
-import diettracker.getUserIdByEmail
-import diettracker.saveCurrentFoodLog
-import diettracker.saveCurrentMeal
-import diettracker.searchFoods
-import diettracker.searchRecipes
+import diettracker.db.repositories.getUserIdByEmail
+import diettracker.services.CaloriesSession
+import diettracker.services.UserSession
+import diettracker.services.addSavedMealToLog
+import diettracker.services.foodLogCustom
+import diettracker.services.foodLogRecipe
+import diettracker.services.foodLogReset
+import diettracker.services.getSavedMeals
+import diettracker.services.saveCurrentFoodLog
+import diettracker.services.saveCurrentMeal
+import diettracker.services.searchFoods
+import diettracker.services.searchRecipes
 import io.ktor.server.pebble.respondTemplate
 import io.ktor.server.response.respondRedirect
 import io.ktor.server.routing.Route
