+++ app/src/main/kotlin/diettracker/routing/MessageRoutes.kt
@@ -1,12 +1,12 @@
 package diettracker.routing
 
-import diettracker.UserSession
-import diettracker.buildNavbarContext
 import diettracker.db.repositories.MessagingRepository
-import diettracker.getClientsForProfessional
-import diettracker.getLinkedProfessionalIdsForClient
-import diettracker.getUserIdByEmail
-import diettracker.getUserRoles
+import diettracker.db.repositories.getClientsForProfessional
+import diettracker.db.repositories.getLinkedProfessionalIdsForClient
+import diettracker.db.repositories.getUserIdByEmail
+import diettracker.db.repositories.getUserRoles
+import diettracker.services.UserSession
+import diettracker.services.buildNavbarContext
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.pebble.respondTemplate
 import io.ktor.server.request.receiveParameters
