+++ app/src/main/kotlin/diettracker/routing/ProfessionalProfileRoutes.kt
@@ -1,10 +1,10 @@
 package diettracker.routing
 
-import diettracker.UserSession
-import diettracker.buildNavbarContext
+import diettracker.db.repositories.getUserIdByEmail
 import diettracker.db.tables.Professionals
 import diettracker.db.tables.Users
-import diettracker.getUserIdByEmail
+import diettracker.services.UserSession
+import diettracker.services.buildNavbarContext
 import io.ktor.server.pebble.respondTemplate
 import io.ktor.server.request.receiveParameters
 import io.ktor.server.response.respondRedirect
@@ -14,7 +14,6 @@ import io.ktor.server.routing.post
 import io.ktor.server.sessions.get
 import io.ktor.server.sessions.sessions
 import org.jetbrains.exposed.v1.core.eq
-import org.jetbrains.exposed.v1.core.innerJoin
 import org.jetbrains.exposed.v1.jdbc.selectAll
 import org.jetbrains.exposed.v1.jdbc.transactions.transaction
 import org.jetbrains.exposed.v1.jdbc.update
