+++ app/src/main/kotlin/diettracker/routing/RecipeRoutes.kt
@@ -1,9 +1,9 @@
 package diettracker.routing
 
-import diettracker.RecipeDatabaseQuery
-import diettracker.UserSession
-import diettracker.buildNavbarContext
-import diettracker.getUserIdByEmail
+import diettracker.db.repositories.RecipeDatabaseQuery
+import diettracker.db.repositories.getUserIdByEmail
+import diettracker.services.UserSession
+import diettracker.services.buildNavbarContext
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.pebble.respondTemplate
 import io.ktor.server.request.receiveParameters
