+++ app/src/main/kotlin/diettracker/routing/StaticRoutes.kt
@@ -1,4 +1,4 @@
-package diettracker
+package diettracker.routing
 
 import io.ktor.server.http.content.staticResources
 import io.ktor.server.routing.Route
