+++ app/src/test/kotlin/diettracker/ClientDietTrendTest.kt
@@ -11,6 +11,7 @@ import diettracker.db.tables.FoodLogItems
 import diettracker.db.tables.FoodLogs
 import diettracker.db.tables.Foods
 import diettracker.db.tables.Users
+import diettracker.services.ClientDietTrend
 import org.jetbrains.exposed.v1.core.eq
 import org.jetbrains.exposed.v1.jdbc.deleteAll
 import org.jetbrains.exposed.v1.jdbc.insert
